requirements.txt
mysql-connector
pudb